Update index.html
Figured out issue with iOS browsers (Safari and Chrome) -- it seems ios renders the colorspace differently when it converts lossy formats such as jpeg, even when the colorspace is specified for the canvas and getImageData()  method. 

This has not been tested with MacOS to see if the issue is the same there.

Using lossless formats such as PNG should eliminate this bug. Will test further