Update needl.class.js
Figured out bug with iOS browsers (Safari and Chrome) -- seems that iOS interprets the colorspace (when getting RGB values from lossy formats such as jpeg) differently than other systems, even when the colorspace is specified in the getContext and getImageData methods. Based on my research this is a common issue with Apple devices and HTML canvas elements. 
I knew this could be a possibility and lossless image formats like PNG should not have this issue.

Will continue to test, but removed a lot of textareaConsole output that was used during debugging.